+++ sample/src/test/java/br/com/concrete/canarinho/test/TesteValidadores.java
@@ -117,11 +117,11 @@ public class TesteValidadores {
 
     @Test
     public void consegueValidarTelefone() {
-        assertThat(Validador.TELFONE.ehValido("1112345678"), is(true));
-        assertThat(Validador.TELFONE.ehValido("11123456789"), is(true));
-        assertThat(Validador.TELFONE.ehValido("(11) 12345-6789"), is(true));
-        assertThat(Validador.TELFONE.ehValido("111234567890"), is(false));
-        assertThat(Validador.TELFONE.ehValido("1112345"), is(false));
+        assertThat(Validador.TELEFONE.ehValido("1112345678"), is(true));
+        assertThat(Validador.TELEFONE.ehValido("11123456789"), is(true));
+        assertThat(Validador.TELEFONE.ehValido("(11) 12345-6789"), is(true));
+        assertThat(Validador.TELEFONE.ehValido("111234567890"), is(false));
+        assertThat(Validador.TELEFONE.ehValido("1112345"), is(false));
     }
 
     @Test